# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientRegisterRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientRegisterResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "User already exists or validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Client registration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request password reset",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid token or password"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reset password",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ruang Diri API",
    "description": "API documentation for Ruang Diri application, all the services belongs to PT. Wong Makmur Sejati",
    "version": "1.0",
    "contact": {
      "name": "Support Team",
      "url": "https://ruangdiri.id",
      "email": "support@ruangdiri.id"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://ruangdiri.id/terms"
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token here"
      }
    },
    "schemas": {
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "securePassword123"
          },
          "rememberMe": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR..."
          },
          "message": {
            "type": "string",
            "example": "Login successful"
          }
        },
        "required": [
          "accessToken",
          "message"
        ]
      },
      "ClientRegisterRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "securePassword123"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Jakarta"
          }
        },
        "required": [
          "email",
          "password",
          "fullName",
          "timezone"
        ]
      },
      "ClientRegisterResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Registration successful"
          },
          "data": {
            "type": "object",
            "example": {
              "id": "123e4567-e89b-12d3-a456-426614174000",
              "email": "user@example.com",
              "fullName": "John Doe",
              "role": "client"
            }
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "ForgotPasswordRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ForgotPasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Password reset email sent"
          }
        },
        "required": [
          "message"
        ]
      },
      "ResetPasswordRequestDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "user@example.com"
          },
          "newPassword": {
            "type": "string",
            "example": "newSecurePassword123"
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "ResetPasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Password reset successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}